+++ src/app/components/product-card/product-card.component.html
@@ -1,7 +1,7 @@
 <div class="card" style="width: 18rem;">
-    <img class="card-img-top" [src]="productInfo.image" alt="Card image cap">
+    <img class="card-img-top" [src]="productInfo.image" alt="Card image cap" [routerLink]="['/product',productInfo.id]">
     <div class="card-body">
-      <h5 class="card-title" [routerLink]="['/product',productInfo.id]">{{productInfo.name}}</h5>
+      <h5 class="card-title" >{{productInfo.name}}</h5>
       <p class="card-text">{{productInfo.description}}</p>
     </div>
     <div class="input-group input-group-sm mb-3">
@@ -16,12 +16,6 @@
         </div>
         <input type="number" class="form-control" aria-label="Sizing example input" [value]="productInfo.quantity" aria-describedby="inputGroup-sizing-sm" disabled>
       </div>
-    <div class="card-body">
-        <button class="btn btn-outline-success" (click)="addToCart(productInfo)">Add to Cart</button>
-    </div>
-
-    <div class="card-body">
-        <button class="btn btn-outline-success" (click)="addToWishCart(productInfo)">Add to Wish List</button>
-
+    
 
   </div>
