+++ src/app/components/product-details/product-details.component.html
@@ -1,1 +1,19 @@
-<p>product-details works!</p>
+<app-navbar></app-navbar>
+<div class="container w-75 mx-auto">
+  <div class="card w-100">
+    <img class="card-img-top" [src]="prodDetail.image" alt="Card image cap">
+    <div class="card-body">
+      <h1 class="text-center display-2">{{prodDetail.name}}</h1>
+      <p class="text-center h3">{{prodDetail.description}}</p>
+      <p type="number" class="text-center" >Price: ${{prodDetail.price}}</p>
+      <p type="number" class="text-center"  >Quantity: {{prodDetail.quantity}}</p>
+    </div>
+
+    <div class="card-footer">
+      <button class="btn btn-outline-success w-100 mb-1" (click)="addToCart(prodDetail)">Add to Cart</button>
+      <button class="btn btn-outline-success w-100 mt-1">Add to Wish List</button>
+    </div>
+
+  </div>
+
+  <!-- (click)="addToWishCart(prodDetail)" -->
