+++ src/app/components/product-details/product-details.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
+import { Product } from 'src/app/models/product';
 import { ProductService } from 'src/app/services/product.service';
 
 @Component({
@@ -9,6 +10,15 @@ import { ProductService } from 'src/app/services/product.service';
 })
 export class ProductDetailsComponent implements OnInit {
   productId: string = '';
+  prodDetail: any;
+  wishCartCount!: number;
+  wishProducts: {
+    wishProduct: Product,
+    wishQuantity: number
+  }[] = [];
+  //wishSubscription!: Subscription;
+  wishTotalPrice!: number;
+
   constructor(private productsService: ProductService, private param: ActivatedRoute) { }
 
   ngOnInit(): void {
@@ -16,18 +26,92 @@ export class ProductDetailsComponent implements OnInit {
     this.getProduct(parseInt(this.productId));
   }
 
+  cartCount!: number;
+  products: {
+    product: Product,
+    quantity: number
+  }[] = [];
+  // subscription!: Subscription;
+  totalPrice: number = 0;
+
   getProduct(id: number) {
     this.productsService.getSingleProduct(id).subscribe({
       next: (response) => {
-        console.log(response);
+        this.prodDetail = response;
       },
-      error: (err)=>{
+      error: (err) => {
         console.log(err);
       },
       complete() {
         console.log("Success Got Product")
       },
     })
-    
+
+  }
+
+  // addToWishCart(product: Product): void {
+  //   let inWishCart = false;
+
+  //       if(this.prodDetail){
+  //         ++this.prodDetail.quantity;
+  //         let wishCart = {
+  //           wishCartCount: this.wishCartCount + 1,
+  //           wishProducts: this.wishProducts,
+  //           wishTotalPrice: this.wishTotalPrice + product.price
+  //         };
+  //         this.productsService.setWishCart(wishCart);
+  //        // inWishCart=true;
+  //         return;
+  //       };
+
+
+  //   if(inWishCart == false){
+  //     let newWishProduct = {
+  //       wishProduct: product,
+  //       wishQuantity: 1
+  //     };
+  //     this.wishProducts.push(newWishProduct);
+  //     let wishCart = {
+  //       wishCartCount: this.wishCartCount + 1,
+  //       wishProducts: this.wishProducts,
+  //       wishTotalPrice: this.wishTotalPrice + product.price
+  //     }
+  //     this.productsService.setWishCart(wishCart);
+  //   }
+  // }
+
+
+  addToCart(product: Product): void {
+
+    // let inCart = false;
+
+    // this.products.forEach(
+    //   (element) => {
+    //     if(element.product == product){
+    //       ++element.quantity;
+    //       let cart = {
+    //         cartCount: this.cartCount + 1,
+    //         products: this.products,
+    //         totalPrice: this.totalPrice + product.price
+    //       };
+    //       this.productsService.setCart(cart);
+    //       inCart=true;
+    //       return;
+    //     };
+    //   }
+    // );
+
+
+    let newProduct = {
+      product: product,
+      quantity: 1
+    };
+    this.products.push(newProduct);
+    let cart = {
+      cartCount: this.cartCount + 1,
+      products: this.products,
+      totalPrice: this.totalPrice + product.price
+    }
+    this.productsService.setCart(cart);
   }
 }
